refactor: extract widgets from formplugin to file
