+++ packages/dm-core-plugins/src/form/widgets/index.ts
@@ -1,7 +1,9 @@
 import { TWidget } from '../types'
+import BlueprintPickerWidget from './BlueprintPickerWidget'
 import CheckboxWidget from './CheckboxWidget'
 import TextWidget from './TextWidget'
 import TextareaWidget from './TextareaWidget'
+import TypeWidget from './TypeWidget'
 
 const widgets: {
   [key: string]: (props: TWidget) => JSX.Element
@@ -9,6 +11,8 @@ const widgets: {
   CheckboxWidget,
   TextWidget,
   TextareaWidget,
+  BlueprintPickerWidget,
+  TypeWidget,
 }
 
 export default widgets
